fix default value loop, let share button copy URL to clipboard
