+++ components/header.tsx
@@ -1,40 +1,16 @@
 "use client";
 
-import * as pako from "pako";
 import {
   NavigationMenu,
   NavigationMenuItem,
-  NavigationMenuLink,
   NavigationMenuList,
-  navigationMenuTriggerStyle,
 } from "@/components/ui/navigation-menu";
 import { Share } from "lucide-react";
-import Link from "next/link";
-import { useContext, useEffect } from "react";
-import { GlobalContext } from "../app/layout";
+import { useCopyToClipboard } from "usehooks-ts";
+import { Button } from "./ui/button";
 
 export function Header() {
-  const { source, setSource } = useContext(GlobalContext);
-
-  useEffect(() => {
-    const input = new TextEncoder().encode(source);
-    const output = pako.deflate(input);
-    const base64 = Buffer.from(output, "binary").toString("base64");
-    window.location.hash = base64;
-    history.pushState(null, "", `#${base64}`);
-  }, [source]);
-
-  useEffect(() => {
-    const handleHashChange = () => {
-      const input = Buffer.from(window.location.hash, "base64");
-      const output = pako.inflate(input);
-      setSource(new TextDecoder().decode(output));
-    };
-    window.addEventListener("hashchange", handleHashChange);
-    return () => {
-      window.removeEventListener("hashchange", handleHashChange);
-    };
-  }, []);
+  const [_, copy] = useCopyToClipboard();
 
   return (
     <header className="dark text-white sticky top-0 z-50 w-full border-b border-border/40 bg-background/95 mb-4 p-2">
@@ -46,11 +22,15 @@ export function Header() {
         <NavigationMenu>
           <NavigationMenuList>
             <NavigationMenuItem>
-              <Link href="/docs" legacyBehavior passHref>
-                <NavigationMenuLink className={navigationMenuTriggerStyle()}>
-                  <Share size={12} className="mr-2" /> Share
-                </NavigationMenuLink>
-              </Link>
+              <Button
+                size="sm"
+                onClick={(e) => {
+                  e.preventDefault();
+                  copy(window.location.href);
+                }}
+              >
+                <Share size={12} className="mr-2" /> Share
+              </Button>
             </NavigationMenuItem>
           </NavigationMenuList>
         </NavigationMenu>
