+++ package.json
@@ -30,6 +30,7 @@
     "@ruby/wasm-wasi": "2.4.1-2024-01-11-a",
     "@svgr/webpack": "^8.1.0",
     "@types/node": "^20.10.6",
+    "@types/pako": "^2.0.3",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "lucide-react": "^0.307.0",
@@ -42,6 +43,7 @@
     "tailwindcss-animate": "^1.0.7",
     "ts-node": "^10.9.2",
     "typescript": "^5.3.3",
+    "usehooks-ts": "^2.9.5",
     "wasmoon": "^1.16.0"
   },
   "devDependencies": {
